fix(staff): match staff_id column type in attendance migration

+++ database/migrations/2026_06_07_192529_create_staff_attendance_table.php
@@ -10,7 +10,7 @@ return new class extends Migration
     {
         Schema::create('staff_attendance', function (Blueprint $table) {
             $table->id('attendance_id');
-            $table->unsignedBigInteger('staff_id');
+            $table->unsignedInteger('staff_id');
             $table->string('staff_name');
             $table->string('role')->nullable();
             $table->string('shift_schedule')->nullable();
